Atualização Psfot em sala.

+++ src/main/java/com/ufcg/psoft/mercadofacil/dto/valid/estabelecimento/CodigoAcessoValidator.java
@@ -3,10 +3,10 @@ package com.ufcg.psoft.mercadofacil.dto.valid.estabelecimento;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class CodigoAcessoValidator implements ConstraintValidator<CodigoAcesso, Long> {
+public class CodigoAcessoValidator implements ConstraintValidator<CodigoAcesso, Integer> {
 
     @Override
-    public boolean isValid(Long value, ConstraintValidatorContext context) {
+    public boolean isValid(Integer value, ConstraintValidatorContext context) {
         if (value == null) {
             return true;
         }
